pyproject.toml
[tool.poetry]
name = "dotaservice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio-tools = "^1.28.1"
poetry = "^1.0.5"

[tool.poetry.dev-dependencies]
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
